Add markdown title cell insertion for ipynb
Adds a function to insert a new first cell or replace the first cell in an
existing ipython notebook with a markdown cell containing text provided.
Can use any ipynb as an input but will not output to any file named
'Metrics Template.ipynb'. This is currently hard coded until a more
elegant solution is devised.